Fix Dataflow breakage introduce with commandline arg parsing changes
